@@ -4,29 +4,29 @@ description: A web-based scriptable cadnano (see https://cadnano.org/ for origin
 #author: David Doty <doty@ucdavis.edu>
 
 environment:
-  sdk: '>=2.9.0 <3.0.0'
+  sdk: '>=2.11.0 <3.0.0'
 
 dependencies:
-  analyzer: '>=0.35.0 <0.40.0'
+  analyzer: '>=0.39.0 <0.42.0'
 #  analyzer: '^1.0.0'
-  spreadsheet_decoder: ^1.2.0
+  spreadsheet_decoder: ^2.0.0
 #  spreadsheet_decoder: ^2.0.0
   smart_dialogs: ^2.1.2
-  dialog: ^0.7.0
+  dialog: ^0.8.0
   meta: ^1.1.7
   js: ^0.6.1+1
-  codemirror: ^0.5.1
-  color: ^2.1.1
+  codemirror: ^0.7.0+5.65.0
+  color: ^3.0.0
   tuple: ^1.0.2
   quiver: ^2.0.5
   test: ^1.14.5
   path: ^1.6.4
-  platform_detect: ^1.3.5
-  dnd: ^1.4.2
+  platform_detect: ^1.0.0
+  dnd: ^2.0.1
 #  dnd: ^2.0.0
 #  redux_thunk: ^0.2.1
-  built_value: ^7.1.0
-  built_collection: ^4.3.0
+  built_value: ^8.1.3
+  built_collection: ^5.1.1
   reselect: ^0.4.0
   redux: ^4.0.0
 #  redux: ^5.0.0
@@ -36,12 +36,12 @@ dependencies:
   over_react: ^4.0.0
 
 dev_dependencies:
-  built_value_generator: ^7.0.0
+  built_value_generator: ^8.0.0
   build_runner: ^1.10.1
   build_test: ^1.0.0
   build_web_compilers: ^2.11.0
   over_react_test: ^2.9.0
-  test_html_builder: ^1.0.0
+  test_html_builder: ^2.0.0
   redux_dev_tools: ^0.5.0
 
 analyzer:
