@@ -72,8 +72,13 @@ GlobalReducer<Design, AppState> design_whole_global_reducer = combineGlobalReduc
 
 // need to operate on Design so we can re-set helix svg coordinates
 Design design_geometry_set_reducer(Design design, AppState state, actions.GeometrySet action) {
-  var new_helices = util.helices_assign_svg(
-      action.geometry, state.ui_state.invert_xy, design.helices.toMap(), design.groups);
+  var new_helices = design.helices.toMap();
+  for (var key in new_helices.keys) {
+    new_helices[key] = new_helices[key].rebuild((b) => b
+      ..geometry.replace(action.geometry)
+    );
+  }
+
   return design.rebuild((b) => b..helices.replace(new_helices)..geometry.replace(action.geometry));
 }
 
