@@ -139,7 +139,7 @@ BuiltSet<int> helix_selections_adjust_reducer(
   var selection_box = action.selection_box;
   var all_helices_in_displayed_group = state.design.helices_in_group(state.ui_state.displayed_group_name);
   List<select.Box> all_bboxes =
-      all_helices_in_displayed_group.values.map((helix) => helix_to_box(helix)).toList();
+      all_helices_in_displayed_group.values.map((helix) => helix_to_box(helix, state.ui_state.invert_y)).toList();
   var selection_box_as_box = select.Box.from_selection_box(selection_box);
   List<Helix> helices_overlapping =
       select.enclosure_list(all_helices_in_displayed_group.values, all_bboxes, selection_box_as_box);
@@ -164,11 +164,11 @@ BuiltSet<int> helix_selections_adjust_reducer(
   return helices_idxs_selected_new.build();
 }
 
-select.Box helix_to_box(Helix helix) {
+select.Box helix_to_box(Helix helix, bool invert_y) {
   //FIXME: this is making boxes that are not far enough apart
   var position3d = helix.position3d;
   num x, y, width, height;
-  var svg_pos = util.position3d_to_side_view_svg(position3d, helix.invert_xy, helix.geometry);
+  var svg_pos = util.position3d_to_side_view_svg(position3d, invert_y, helix.geometry);
   x = svg_pos.x - helix.geometry.helix_radius_svg;
   y = svg_pos.y - helix.geometry.helix_radius_svg;
   height = width = helix.geometry.helix_radius_svg * 2.0;
