@@ -74,7 +74,7 @@ abstract class AppUIStateStorables
 
   String get loaded_script_filename;
 
-  bool get invert_yz;
+  bool get invert_y;
 
   bool get warn_on_exit_if_unsaved;
 
@@ -130,7 +130,7 @@ abstract class AppUIStateStorables
     b.display_major_tick_widths_all_helices = false;
     b.loaded_filename = default_filename();
     b.loaded_script_filename = default_script_filename();
-    b.invert_yz = false;
+    b.invert_y = false;
     b.warn_on_exit_if_unsaved = true;
     b.show_helix_circles_main_view = true;
     b.show_edit_mode_menu = true;
