@@ -26,20 +26,26 @@ abstract class DomainsMove with BuiltJsonSerializable implements Built<DomainsMo
   factory DomainsMove(
       {BuiltList<Domain> domains_moving,
       BuiltList<Domain> all_domains,
-        BuiltList<Strand> strands_with_domains_moving,
+      BuiltList<Strand> strands_with_domains_moving,
       BuiltMap<int, Helix> helices,
       BuiltMap<String, HelixGroup> groups,
+      BuiltMap<int, int> original_helices_view_order_inverse,
       Address original_address,
       bool copy = false,
       bool keep_color = true}) {
-    var domains_fixed =
-        copy ? all_domains : [for (var domain in all_domains) if (!domains_moving.contains(domain)) domain];
+    var domains_fixed = copy
+        ? all_domains
+        : [
+            for (var domain in all_domains)
+              if (!domains_moving.contains(domain)) domain
+          ];
     return DomainsMove.from((b) => b
       ..domains_moving.replace(domains_moving)
       ..domains_fixed.replace(domains_fixed)
       ..strands_with_domains_moving.replace(strands_with_domains_moving)
       ..helices.replace(helices)
       ..groups.replace(groups)
+      ..original_helices_view_order_inverse.replace(original_helices_view_order_inverse)
       ..original_address.replace(original_address)
       ..current_address.replace(original_address)
       ..copy = copy
@@ -56,6 +62,10 @@ abstract class DomainsMove with BuiltJsonSerializable implements Built<DomainsMo
   BuiltMap<String, HelixGroup> get groups;
 
   BuiltList<Strand> get strands_with_domains_moving;
+
+  // Since copied Domains may come from a different Design with different groups, we need to
+  // store this to know how to position them in new HelixGroups.
+  BuiltMap<int, int> get original_helices_view_order_inverse;
 
   Address get original_address;
 
