@@ -1,3 +1,5 @@
+import 'dart:html';
+
 import 'package:over_react/over_react.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:scadnano/src/state/group.dart';
@@ -18,6 +20,7 @@ mixin DesignMainDNAMismatchesProps on UiProps {
   Design design;
   bool only_display_selected_helices;
   BuiltSet<int> side_selected_helix_idxs;
+  BuiltMap<int, num> helix_idx_to_svg_position_y_map;
 }
 
 class DesignMainDNAMismatchesComponent extends UiComponent2<DesignMainDNAMismatchesProps> with PureComponent {
@@ -38,7 +41,7 @@ class DesignMainDNAMismatchesComponent extends UiComponent2<DesignMainDNAMismatc
         for (Mismatch mismatch in mismatches) {
           var helix = props.design.helices[domain.helix];
           if (!props.only_display_selected_helices || props.side_selected_helix_idxs.contains(helix.idx)) {
-            var base_svg_pos = helix.svg_base_pos(mismatch.offset, domain.forward);
+            var base_svg_pos = helix.svg_base_pos(mismatch.offset, domain.forward, props.helix_idx_to_svg_position_y_map[helix.idx]);
             // For now, if there is a mismatch in an insertion we simply display it for the whole insertion,
             // not for a specific base. We maintain React keys to agree on any mismatches in the same
             // insertion, and we only render one of them.
