@@ -1,3 +1,5 @@
+import 'dart:html';
+
 import 'package:built_collection/built_collection.dart';
 import 'package:color/color.dart';
 import 'package:over_react/over_react.dart';
@@ -16,7 +18,7 @@ part 'design_main_domain_moving.over_react.g.dart';
 UiFactory<DesignMainDomainMovingProps> DesignMainDomainMoving = _$DesignMainDomainMoving;
 
 mixin DesignMainDomainMovingPropsMixin on UiProps {
-  Domain domain;
+  Domain domain_moved;
   Color color;
   HelixGroup original_group;
   HelixGroup current_group;
@@ -29,6 +31,7 @@ mixin DesignMainDomainMovingPropsMixin on UiProps {
   BuiltMap<int, Helix> helices;
   BuiltMap<String, HelixGroup> groups;
   Geometry geometry;
+  num domain_helix_svg_position_y;
 }
 
 class DesignMainDomainMovingProps = UiProps
@@ -38,27 +41,22 @@ class DesignMainDomainMovingComponent extends UiComponent2<DesignMainDomainMovin
     with PureComponent, TransformByHelixGroup<DesignMainDomainMovingProps> {
   @override
   render() {
-    Domain domain_moved = move_domain(
-      domain: props.domain,
-      original_group: props.original_group,
-      current_group: props.current_group,
-      delta_view_order: props.delta_view_order,
-      delta_offset: props.delta_offset,
-      delta_forward: props.delta_forward,
-      set_first_last_false: true, // don't want to display 5'/3' ends while moving
-    );
-
-    Helix helix = props.helices[domain_moved.helix];
-    var start_svg = helix.svg_base_pos(domain_moved.offset_5p, domain_moved.forward);
-    var end_svg = helix.svg_base_pos(domain_moved.offset_3p, domain_moved.forward);
+    Helix helix = props.helices[props.domain_moved.helix];
+    var start_svg = helix.svg_base_pos(
+        props.domain_moved.offset_5p, props.domain_moved.forward, props.domain_helix_svg_position_y);
+    var end_svg = helix.svg_base_pos(
+        props.domain_moved.offset_3p, props.domain_moved.forward, props.domain_helix_svg_position_y);
 
     var classname = constants.css_selector_domain_moving;
     if (!props.allowable) {
       classname += ' ' + constants.css_selector_disallowed;
     }
 
+    var hex_color = props.color.toHexColor();
+    var hex_color_css = hex_color.toCssString();
+
     return (Dom.line()
-      ..stroke = props.color.toHexColor().toCssString()
+      ..stroke = hex_color_css
       ..transform = transform_of_helix(helix.idx)
       ..x1 = '${start_svg.x}'
       ..y1 = '${start_svg.y}'
