@@ -1,3 +1,5 @@
+import 'dart:html';
+
 import 'package:over_react/over_react.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:tuple/tuple.dart';
@@ -18,6 +20,7 @@ mixin DesignMainDomainNameMismatchesProps on UiProps {
   Design design;
   bool only_display_selected_helices;
   BuiltSet<int> side_selected_helix_idxs;
+  BuiltMap<int, Point<num>> helix_idx_to_svg_position_map;
 }
 
 class DesignMainDomainNameMismatchesComponent extends UiComponent2<DesignMainDomainNameMismatchesProps>
@@ -46,7 +49,7 @@ class DesignMainDomainNameMismatchesComponent extends UiComponent2<DesignMainDom
         // draw mismatch stars at midpoint of overlap of domains
         int mid = (overlap.item1 + overlap.item2) ~/ 2;
         for (Domain domain in [forward_domain, reverse_domain]) {
-          var base_svg_pos = helix.svg_base_pos(mid, domain.forward);
+          var base_svg_pos = helix.svg_base_pos(mid, domain.forward, props.helix_idx_to_svg_position_map[helix.idx].y);
           String key = '${domain.helix};${domain.forward};${domain.start};${mid};${domain.end}';
           var mismatch_component = (DesignMainMismatch()
             ..base_svg_pos = base_svg_pos
