@@ -1,3 +1,4 @@
+import 'dart:html';
 import 'dart:math' as Math;
 import 'package:over_react/over_react.dart';
 import 'package:built_collection/built_collection.dart';
@@ -27,6 +28,7 @@ mixin DesignMainSliceBarProps on UiProps {
   BuiltMap<int, Helix> helices;
   bool only_display_selected_helices;
   Geometry geometry;
+  BuiltMap<int, Point<num>> helix_idx_to_svg_position_map;
 }
 
 class DesignMainSliceBarComponent extends UiComponent2<DesignMainSliceBarProps> with PureComponent {
@@ -42,7 +44,7 @@ class DesignMainSliceBarComponent extends UiComponent2<DesignMainSliceBarProps> 
       bool only_display_selected_helices = props.only_display_selected_helices;
       if (only_display_selected_helices && side_selected_helix_idxs.contains(helix.idx) ||
           !only_display_selected_helices) {
-        var y = helix.svg_position.y;
+        var y = props.helix_idx_to_svg_position_map[helix_idx].y;
         displayed_helices_max_y = Math.max(displayed_helices_max_y, y);
         displayed_helices_min_y = Math.min(displayed_helices_min_y, y);
       }
@@ -57,7 +59,7 @@ class DesignMainSliceBarComponent extends UiComponent2<DesignMainSliceBarProps> 
     var slice_bar_svg_width = geometry.base_width_svg;
     var slice_bar_svg_height = displayed_helices_max_y - displayed_helices_min_y + geometry.helix_diameter_svg;
     var helix = props.helices[helix_idxs_in_group.first];
-    var slice_bar_svg_base_center_pos = helix.svg_base_pos(props.slice_bar_offset, true);
+    var slice_bar_svg_base_center_pos = helix.svg_base_pos(props.slice_bar_offset, true, props.helix_idx_to_svg_position_map[helix_idxs_in_group.first].y);
 
     var slice_bar_x = slice_bar_svg_base_center_pos.x - geometry.base_width_svg / 2;
     var slice_bar_y = displayed_helices_min_y - geometry.helix_radius_svg + geometry.base_height_svg;
