@@ -46,6 +46,8 @@ mixin DesignMainLoopoutPropsMixin on UiProps {
   BuiltMap<int, Helix> helices;
   BuiltMap<String, HelixGroup> groups;
   Geometry geometry;
+  num prev_helix_svg_position_y;
+  num next_helix_svg_position_y;
 }
 
 class DesignMainLoopoutProps = UiProps with DesignMainLoopoutPropsMixin, TransformByHelixGroupPropsMixin;
@@ -86,7 +88,7 @@ class DesignMainLoopoutComponent extends UiStatefulComponent2<DesignMainLoopoutP
 
     if (within_group) {
       path_description = loopout_path_description_within_group(props.prev_helix, props.next_helix,
-          props.prev_domain, props.next_domain, props.loopout, true, props.show_domain_names);
+          props.prev_domain, props.next_domain, props.loopout, true, props.show_domain_names, props.prev_helix_svg_position_y, props.next_helix_svg_position_y);
     } else {
       path_description = loopout_path_description_between_groups();
     }
@@ -213,8 +215,8 @@ class DesignMainLoopoutComponent extends UiStatefulComponent2<DesignMainLoopoutP
     var prev_group = props.groups[props.prev_helix.group];
     var next_group = props.groups[props.next_helix.group];
 
-    var prev_svg_untransformed = props.prev_helix.svg_base_pos(prev_offset, props.prev_domain.forward);
-    var next_svg_untransformed = props.next_helix.svg_base_pos(next_offset, props.next_domain.forward);
+    var prev_svg_untransformed = props.prev_helix.svg_base_pos(prev_offset, props.prev_domain.forward, props.prev_helix_svg_position_y);
+    var next_svg_untransformed = props.next_helix.svg_base_pos(next_offset, props.next_domain.forward, props.next_helix_svg_position_y);
 
     var prev_svg = prev_group.transform_point_main_view(prev_svg_untransformed, props.geometry);
     var next_svg = next_group.transform_point_main_view(next_svg_untransformed, props.geometry);
@@ -248,7 +250,7 @@ class DesignMainLoopoutComponent extends UiStatefulComponent2<DesignMainLoopoutP
 // When drawing a normal loopout this is needed, but when drawing a moving strand, where all path commands
 // are concatenated together, it is not needed.
 String loopout_path_description_within_group(Helix prev_helix, Helix next_helix, Domain prev_domain,
-    Domain next_domain, Loopout loopout, bool include_start_M, bool show_loopout_labels) {
+    Domain next_domain, Loopout loopout, bool include_start_M, bool show_loopout_labels, num prev_helix_svg_position_y, num next_helix_svg_position_y) {
   Helix top_helix = prev_helix;
   Helix bot_helix = next_helix;
   Geometry geometry = top_helix.geometry;
@@ -258,8 +260,8 @@ String loopout_path_description_within_group(Helix prev_helix, Helix next_helix,
   bool same_helix = top_helix.idx == bot_helix.idx;
   if (same_helix && top_dom.forward == bot_dom.forward) {
     return loopout_path_description_same_helix_same_direction(
-        loopout, prev_helix, prev_domain, next_domain, include_start_M, show_loopout_labels);
-  } else if (top_helix.svg_position.y > bot_helix.svg_position.y || same_helix && !top_dom.forward) {
+        loopout, prev_helix, prev_domain, next_domain, include_start_M, show_loopout_labels, prev_helix_svg_position_y);
+  } else if (prev_helix_svg_position_y > next_helix_svg_position_y || same_helix && !top_dom.forward) {
     // helices are same if same_helix is true, but we'd still like top_dom/bot_dom to reflect which
     // is on top, i.e. that of the forward domain
     top_helix = next_helix;
@@ -274,8 +276,8 @@ String loopout_path_description_within_group(Helix prev_helix, Helix next_helix,
   int prev_offset = top_dom_is_prev ? top_offset : bot_offset;
   int next_offset = top_dom_is_prev ? bot_offset : top_offset;
 
-  var prev_svg = prev_helix.svg_base_pos(prev_offset, prev_domain.forward);
-  var next_svg = next_helix.svg_base_pos(next_offset, next_domain.forward);
+  var prev_svg = prev_helix.svg_base_pos(prev_offset, prev_domain.forward, prev_helix_svg_position_y);
+  var next_svg = next_helix.svg_base_pos(next_offset, next_domain.forward, next_helix_svg_position_y);
 
   var w, h;
 
@@ -321,13 +323,13 @@ String loopout_path_description_within_group(Helix prev_helix, Helix next_helix,
 }
 
 String loopout_path_description_same_helix_same_direction(Loopout loopout, Helix helix, Domain prev_domain,
-    Domain next_domain, bool include_start_m, bool show_loopout_labels) {
+    Domain next_domain, bool include_start_m, bool show_loopout_labels, num helix_svg_position_y) {
   Geometry geometry = helix.geometry;
   int prev_offset = prev_domain.offset_3p;
   int next_offset = next_domain.offset_5p;
   bool forward = prev_domain.forward;
-  var prev_svg = helix.svg_base_pos(prev_offset, forward);
-  var next_svg = helix.svg_base_pos(next_offset, forward);
+  var prev_svg = helix.svg_base_pos(prev_offset, forward, helix_svg_position_y);
+  var next_svg = helix.svg_base_pos(next_offset, forward, helix_svg_position_y);
 
   int left_offset = prev_offset;
   int right_offset = next_offset;
