@@ -1,7 +1,10 @@
+import 'dart:html';
+
 import 'package:over_react/over_react.dart';
 import 'package:built_collection/built_collection.dart';
 import 'package:over_react/over_react_redux.dart';
 import 'package:react/react_client/react_interop.dart';
+import 'package:scadnano/src/state/dna_assign_options.dart';
 
 import '../state/group.dart';
 import '../state/geometry.dart';
@@ -24,8 +27,7 @@ UiFactory<DesignMainStrandsProps> ConnectedDesignMainStrands =
     ..selectables_store = state.ui_state.selectables_store
     ..drawing_potential_crossover = state.ui_state.potential_crossover_is_drawing
     ..moving_dna_ends = state.ui_state.dna_ends_are_moving
-    ..assign_complement_to_bound_strands_default = state.ui_state.assign_complement_to_bound_strands_default
-    ..warn_on_change_strand_dna_assign_default = state.ui_state.warn_on_change_strand_dna_assign_default
+    ..dna_assign_options = state.ui_state.dna_assign_options
     ..only_display_selected_helices = state.ui_state.only_display_selected_helices
     ..show_dna = state.ui_state.show_dna
     ..show_modifications = state.ui_state.show_modifications
@@ -35,6 +37,7 @@ UiFactory<DesignMainStrandsProps> ConnectedDesignMainStrands =
     ..show_strand_names = state.ui_state.show_strand_names
     ..domain_name_font_size = state.ui_state.domain_name_font_size
     ..strand_name_font_size = state.ui_state.strand_name_font_size
+    ..helix_idx_to_svg_position_map = state.helix_idx_to_svg_position_map
     ..geometry = state.design.geometry;
 })(DesignMainStrands);
 
@@ -55,11 +58,11 @@ mixin DesignMainStrandsProps on UiProps {
   num modification_font_size;
   bool drawing_potential_crossover;
   bool moving_dna_ends;
-  bool assign_complement_to_bound_strands_default;
-  bool warn_on_change_strand_dna_assign_default;
+  DNAAssignOptions dna_assign_options;
   bool only_display_selected_helices;
   bool modification_display_connector;
   Geometry geometry;
+  BuiltMap<int, Point<num>> helix_idx_to_svg_position_map;
 }
 
 class DesignMainStrandsComponent extends UiComponent2<DesignMainStrandsProps> with PureComponent {
@@ -99,8 +102,7 @@ class DesignMainStrandsComponent extends UiComponent2<DesignMainStrandsProps> wi
         ..selected_modifications_in_strand = selected_modifications_in_strand
         ..drawing_potential_crossover = props.drawing_potential_crossover
         ..moving_dna_ends = props.moving_dna_ends
-        ..assign_complement_to_bound_strands_default = props.assign_complement_to_bound_strands_default
-        ..warn_on_change_strand_dna_assign_default = props.warn_on_change_strand_dna_assign_default
+        ..dna_assign_options = props.dna_assign_options
         ..only_display_selected_helices = props.only_display_selected_helices
         ..show_dna = props.show_dna
         ..show_modifications = props.show_modifications
@@ -111,6 +113,7 @@ class DesignMainStrandsComponent extends UiComponent2<DesignMainStrandsProps> wi
         ..modification_font_size = props.modification_font_size
         ..modification_display_connector = props.modification_display_connector
         ..geometry = props.geometry
+        ..helix_idx_to_svg_position_map = props.helix_idx_to_svg_position_map
         ..key = strand.toString())());
     }
 
