@@ -103,7 +103,7 @@ class DesignSideHelixComponent extends UiComponent2<DesignSideHelixProps> with P
     }
 
     Position3D pos3d = props.helix.position3d;
-    Point<num> center = util.position3d_to_side_view_svg(pos3d, props.helix.invert_xy, props.helix.geometry);
+    Point<num> center = util.position3d_to_side_view_svg(pos3d, props.invert_y, props.helix.geometry);
 
     return (Dom.g()
       ..transform = 'translate(${center.x} ${center.y})'
