@@ -221,9 +221,7 @@ main() {
     var geometry = state.design.geometry;
     final correct_helix =
         new Helix(grid_position: grid_position, idx: 0, grid: Grid.square, geometry: geometry);
-    var correct_helices =
-        util.helices_assign_svg(geometry, false, {correct_helix.idx: correct_helix}, state.design.groups);
-    expect(state.design.helices, BuiltMap<int, Helix>(correct_helices));
+    expect(state.design.helices, BuiltMap<int, Helix>({correct_helix.idx: correct_helix}));
   });
 
   test('should_remove_helix_in_response_to_HelixRemove', () {
@@ -1761,15 +1759,20 @@ main() {
   Design two_helices_join_inner_strands = Design.from_json(jsonDecode(two_helices_join_inner_strands_json));
   test('pencil should connect a 3p end to a 5p end', () {
     AppState state = app_state_from_design(two_helices_design);
+    Map<int, Point<num>> svg_position_map = util.helices_assign_svg(two_helices_design.geometry,
+        state.ui_state.invert_y, two_helices_design.helices, two_helices_design.groups);
 
     Strand h0_reverse_strand = two_helices_design.strands[1];
     Strand h1_forward_strand = two_helices_design.strands[2];
     Helix h0 = two_helices_design.helices[0];
-    Point<num> start_point = h0.svg_base_pos(0, false); // 3p end is 0 offset and forward is false.
+    Point<num> start_point =
+        h0.svg_base_pos(0, false, svg_position_map[0].y); // 3p end is 0 offset and forward is false.
     PotentialCrossover helix_0_3p_end_potential_crossover = PotentialCrossover(
-      helix_idx: 0,
-      forward: false,
-      offset: 0,
+      address: Address(
+        helix_idx: 0,
+        forward: false,
+        offset: 0,
+      ),
       color: h0_reverse_strand.color.toHexColor().toCssString(),
       dna_end_first_click: h0_reverse_strand.dnaend_3p,
       start_point: start_point,
@@ -1788,15 +1791,20 @@ main() {
   });
   test('pencil should connect a 5p end to a 3p end', () {
     AppState state = app_state_from_design(two_helices_design);
+    Map<int, Point<num>> svg_position_map = util.helices_assign_svg(two_helices_design.geometry,
+        state.ui_state.invert_y, two_helices_design.helices, two_helices_design.groups);
 
     Strand h0_reverse_strand = two_helices_design.strands[1];
     Strand h1_forward_strand = two_helices_design.strands[2];
     Helix h1 = two_helices_design.helices[1];
-    Point<num> start_point = h1.svg_base_pos(0, true); // 5p end is 0 offset and forward is true.
+    Point<num> start_point =
+        h1.svg_base_pos(0, true, svg_position_map[1].y); // 5p end is 0 offset and forward is true.
     PotentialCrossover helix_1_5p_end_potential_crossover = PotentialCrossover(
-      helix_idx: 1,
-      forward: true,
-      offset: 0,
+      address: Address(
+        helix_idx: 1,
+        forward: true,
+        offset: 0,
+      ),
       color: h1_forward_strand.color.toHexColor().toCssString(),
       dna_end_first_click: h1_forward_strand.dnaend_5p,
       start_point: start_point,
@@ -1844,15 +1852,11 @@ main() {
     AppState final_state = app_state_reducer(original_state, HelixRemove(0));
     Design final_design = final_state.design;
 
-    Geometry geometry = two_helices_design.geometry;
-
     Helix helix1 = two_helices_design.helices[1];
-    num svg_y_helix_1 = helix1.grid_position.v * geometry.distance_between_helices_svg;
 
-    Helix new_helix1 = helix1.rebuild((b) => b..svg_position_ = Point(0, svg_y_helix_1));
     BuiltList<Strand> new_strands = two_helices_design.strands.rebuild((b) => b..removeRange(0, 2));
     Design expected_design =
-        two_helices_design.rebuild((b) => b..helices.replace({1: new_helix1})..strands.replace(new_strands));
+        two_helices_design.rebuild((b) => b..helices.replace({1: helix1})..strands.replace(new_strands));
 
     expect_design_equal(final_design, expected_design);
   });
@@ -4440,10 +4444,12 @@ main() {
     //
     // 1 [------------------->
     //   <-------------------]
+    Map<int, Point<num>> svg_position_map = util.helices_assign_svg(
+        two_helices_design.geometry, false, two_helices_design.helices, two_helices_design.groups);
 
     DNAEnd dnaEnd = two_helices_design.strands.first.dnaend_5p;
     Helix helix0 = two_helices_design.helices.values.first;
-    Point<num> start_point = helix0.svg_base_pos(0, true);
+    Point<num> start_point = helix0.svg_base_pos(0, true, svg_position_map[helix0.idx].y);
 
     // The two states of the two store's reducers we want to test:
     AppState state = app_state_from_design(two_helices_design);
@@ -4459,9 +4465,11 @@ main() {
     // 1 [------------------->
     //   <-------------------]
     PotentialCrossover potentialCrossover = PotentialCrossover(
-      helix_idx: 0,
-      offset: 0,
-      forward: true,
+      address: Address(
+        helix_idx: 0,
+        offset: 0,
+        forward: true,
+      ),
       color: '#000',
       dna_end_first_click: dnaEnd,
       start_point: start_point,
@@ -5094,7 +5102,7 @@ main() {
       expect(state.ui_state.strands_move, expected_strands_move);
     });
   });
-  
+
   group('insertion/deletion', () {
     //TODO: add some unit tests that test this when DNA sequences are on the domains, and both domains
     // having the deletion/insertion added are part of the same strand (e.g., a hairpin);
@@ -5421,17 +5429,15 @@ main() {
       Helix new_helix0 = no_grid_two_helices_design.helices.values.first.rebuild((b) => b
         ..grid = grid
         ..position_ = null
-        ..grid_position.replace(expected_grid_position0));
+        ..grid_position.replace(expected_grid_position0)
+        ..geometry.replace(no_grid_two_helices_design.geometry));
       Helix new_helix1 = no_grid_two_helices_design.helices.values.last.rebuild((b) => b
         ..grid = grid
         ..position_ = null
-        ..grid_position.replace(expected_grid_position1));
+        ..grid_position.replace(expected_grid_position1)
+        ..geometry.replace(no_grid_two_helices_design.geometry));
 
       Map<int, Helix> new_helices = {0: new_helix0, 1: new_helix1};
-      // need to reassign SVG here since original design had positive x Position3D, which means
-      // positive svi_position.x
-      new_helices = util.helices_assign_svg(
-          no_grid_two_helices_design.geometry, false, new_helices, state.design.groups);
 
       Design expected_design = no_grid_two_helices_design.rebuild((b) => b..helices.replace(new_helices));
       expected_design = expected_design.set_grid(grid);
@@ -5462,43 +5468,17 @@ main() {
     expect_design_equal(state.design, expected_design);
   });
 
-  test('helix_svg_position_from_position', () {
-    AppState state = app_state_from_design(no_grid_two_helices_design);
-    // helix 0 old position: Position3D(x: 10, y: 60, z: 30);
-    // helix 1 old position: Position3D(x: 20, y: 80, z: 50);
-    Geometry geometry = no_grid_two_helices_design.geometry;
-    Helix helix0 = no_grid_two_helices_design.helices[0];
-    Helix helix1 = no_grid_two_helices_design.helices[1];
-    Point<num> svg_position0 = Point<num>(30, 60) * geometry.nm_to_svg_pixels;
-    Point<num> svg_position1 = Point<num>(50 * geometry.nm_to_svg_pixels,
-        svg_position0.y + util.norm_l2(20 - 10, 80 - 60) * geometry.nm_to_svg_pixels);
-
-    Helix expected_helix0 = helix0.rebuild((b) => b..svg_position_ = svg_position0);
-    Helix expected_helix1 = helix1.rebuild((b) => b..svg_position_ = svg_position1);
-    var expected_helices = {0: expected_helix0, 1: expected_helix1};
-
-    Design expected_design = no_grid_two_helices_design.rebuild((b) => b.helices.replace(expected_helices));
-    expect(state.design, expected_design);
-  });
-
   test('HelixPositionSet', () {
     AppState state = app_state_from_design(no_grid_two_helices_design);
     // helix 0 old position: Position3D(x: 10, y: 60, z: 30);
     // helix 0 new position: Position3D(x: 40, y: 30, z: 130);
     // helix 1 old position: Position3D(x: 20, y: 80, z: 50);
-    Geometry geometry = no_grid_two_helices_design.geometry;
     Helix helix0 = no_grid_two_helices_design.helices[0];
     Helix helix1 = no_grid_two_helices_design.helices[1];
     Position3D new_position0 = Position3D(x: 40, y: 30, z: 130);
-    Point<num> svg_position0 = Point<num>(130, 30) * geometry.nm_to_svg_pixels;
-    Point<num> svg_position1 = Point<num>(50 * geometry.nm_to_svg_pixels,
-        svg_position0.y + util.norm_l2(20 - 40, 80 - 30) * geometry.nm_to_svg_pixels);
 
-    Helix expected_helix0 = helix0.rebuild((b) => b
-      ..position_.replace(new_position0)
-      ..svg_position_ = svg_position0);
-    Helix expected_helix1 = helix1.rebuild((b) => b..svg_position_ = svg_position1);
-    var expected_helices = {0: expected_helix0, 1: expected_helix1};
+    Helix expected_helix0 = helix0.rebuild((b) => b..position_.replace(new_position0));
+    var expected_helices = {0: expected_helix0, 1: helix1};
 
     state = app_state_reducer(state, HelixPositionSet(helix_idx: helix0.idx, position: new_position0));
     Design expected_design = no_grid_two_helices_design.rebuild((b) => b.helices.replace(expected_helices));
@@ -5511,21 +5491,13 @@ main() {
     // helix 1 old position: Position3D(x: 20, y: 80, z: 50);
     // helix 0 new position: Position3D(x: 200, y: 160, z: 10);
     // helix 1 new position: Position3D(x: 300, y: 280, z: 500);
-    Geometry geometry = no_grid_two_helices_design.geometry;
     Helix helix0 = no_grid_two_helices_design.helices[0];
     Helix helix1 = no_grid_two_helices_design.helices[1];
     Position3D position0 = Position3D(x: 200, y: 160, z: 10);
     Position3D position1 = Position3D(x: 300, y: 280, z: 500);
-    Point<num> svg_position0 = Point<num>(10, 160) * geometry.nm_to_svg_pixels;
-    Point<num> svg_position1 = Point<num>(500 * geometry.nm_to_svg_pixels,
-        svg_position0.y + util.norm_l2(300 - 200, 280 - 160) * geometry.nm_to_svg_pixels);
 
-    Helix expected_helix0 = helix0.rebuild((b) => b
-      ..position_.replace(position0)
-      ..svg_position_ = svg_position0);
-    Helix expected_helix1 = helix1.rebuild((b) => b
-      ..position_.replace(position1)
-      ..svg_position_ = svg_position1);
+    Helix expected_helix0 = helix0.rebuild((b) => b..position_.replace(position0));
+    Helix expected_helix1 = helix1.rebuild((b) => b..position_.replace(position1));
 
     var batch_action = BatchAction([
       HelixPositionSet(helix_idx: helix0.idx, position: position0),
@@ -7532,21 +7504,65 @@ main() {
   test('adjusting_helices_view_order_should_update_svg_position', () {
     Helix helix0 = Helix(idx: 0, grid_position: GridPosition(0, 0), group: "foo");
     Helix helix1 = Helix(idx: 1, grid_position: GridPosition(0, 1), group: "foo");
-    HelixGroup group = HelixGroup(helices_view_order: [0,1], grid: Grid.square);
+    HelixGroup group = HelixGroup(helices_view_order: [0, 1], grid: Grid.square);
     Design design = Design(helices: [helix0, helix1], groups: {"foo": group});
     helix0 = design.helices[0];
     helix1 = design.helices[1];
     AppState state = app_state_from_design(design);
     // Point<num> original_helix0_svg_position = state.design.helices[0].svg_position;
-    Point<num> original_helix1_svg_position = state.design.helices[1].svg_position;
+    Point<num> original_helix1_svg_position = state.helix_idx_to_svg_position_map[1];
 
-    HelixGroup new_group = HelixGroup(helices_view_order: [1,0], grid: Grid.square);
-    AppState new_state = app_state_reducer(state, GroupChange(old_name: "foo", new_name: "bar", new_group: new_group));
+    HelixGroup new_group = HelixGroup(helices_view_order: [1, 0], grid: Grid.square);
+    AppState new_state =
+        app_state_reducer(state, GroupChange(old_name: "foo", new_name: "bar", new_group: new_group));
 
     // New svg position y coordinate should have changed
-    expect(new_state.design.helices[1].svg_position.y, closeTo(original_helix1_svg_position.y, 0.001));
+    expect(new_state.helix_idx_to_svg_position_map[1].y, closeTo(original_helix1_svg_position.y, 0.001));
     var offset = (helix1.position3d.y - helix0.position3d.y) * design.geometry.nm_to_svg_pixels;
-    expect(new_state.design.helices[0].svg_position.y, closeTo(original_helix1_svg_position.y + offset, 0.001));
+    expect(new_state.helix_idx_to_svg_position_map[0].y,
+        closeTo(original_helix1_svg_position.y + offset, 0.001));
+  });
+
+  // See https://github.com/UC-Davis-molecular-computing/scadnano/pull/703#issuecomment-1003738517
+  // Test Description:
+  //
+  // Setup:
+  //     Helix positions:
+  //             0     1
+  //        0 helix0 helix2
+  //        1 helix1
+  //
+  //     - helix0 and helix2 are selected, helix1 is not selected
+  //     - only_display_selected_helices is false, so all three helices
+  //       appears in the main view
+  //
+  // Action: SetOnlyDisplaySelectedHelices
+  //
+  // Expected Result:
+  //     - helix0 stays in same svg position, helix2 now should have same svg
+  //       position that helix1 initially had, since helix1 and helix2 are both
+  //       the same distance from helix0
+  test('display_only_selected_helices_should_reassign_svg', () {
+    // Setup:
+    Helix helix0 = Helix(idx: 0, grid_position: GridPosition(0, 0));
+    Helix helix1 = Helix(idx: 1, grid_position: GridPosition(0, 1));
+    Helix helix2 = Helix(idx: 2, grid_position: GridPosition(1, 0));
+    Design design = Design(helices: [helix0, helix1, helix2], grid: Grid.square);
+    AppState state = app_state_from_design(design).rebuild((b) => b
+      ..ui_state.storables.side_selected_helix_idxs.replace([0, 2])
+      ..ui_state.storables.only_display_selected_helices = false);
+
+    // Action
+    AppState new_state = app_state_reducer(state, SetOnlyDisplaySelectedHelices(true));
+
+    // Verify Expected Result:
+    Point<num> helix0_svg = state.helix_idx_to_svg_position_map[0];
+    Point<num> helix1_svg = state.helix_idx_to_svg_position_map[1];
+    Point<num> new_helix0_svg = new_state.helix_idx_to_svg_position_map[0];
+    Point<num> new_helix2_svg = new_state.helix_idx_to_svg_position_map[2];
+
+    expect(new_helix0_svg.y, closeTo(helix0_svg.y, 0.001));
+    expect(new_helix2_svg.y, closeTo(helix1_svg.y, 0.001));
   });
 }
 
